Tests/LibWeb/Text/expected/input-number-typing-fractional.txt

After '1': value="1"
After '.': value=""
After '5': value="5"
